EnergyMonitoringSystem/backend/main.py
#!/usr/bin/env python3
"""
PAC3220 Prepaid Energy Monitoring System - FastAPI Backend
Main application entry point with all API routes.
"""

import os
import sys
from pathlib import Path
from fastapi import FastAPI, Request, HTTPException, Depends
from fastapi.middleware.cors import CORSMiddleware
from starlette.middleware.base import BaseHTTPMiddleware
from time import time
from datetime import datetime
from fastapi.staticfiles import StaticFiles
from fastapi.responses import FileResponse, JSONResponse
from fastapi import WebSocket, WebSocketDisconnect
import uvicorn

# Add current directory to path for imports
sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))

# Add backend directory and project root to sys.path so absolute 'backend' imports work
backend_dir = os.path.dirname(os.path.abspath(__file__))
project_root = os.path.dirname(backend_dir)
if backend_dir not in sys.path:
    sys.path.insert(0, backend_dir)
if project_root not in sys.path:
    sys.path.insert(0, project_root)

from backend.api.routes_auth import router as auth_router
from backend.api.routes_auth import login as auth_login
from backend.api.routes_auth import LoginRequest, TokenResponse
from backend.api.routes_auth import create_jwt_token
from backend.api.routes_auth import get_current_user
from backend.api.routes_admin import router as admin_router
from backend.api.routes_dashboard import router as dashboard_router
from backend.api.routes_devices import router as devices_router
from backend.api.routes_readings import router as readings_router
from backend.api.routes_tariffs import router as tariffs_router
from backend.api.routes_do_control import router as do_router
from backend.websocket_manager import ws_manager
from backend.dal.database import db_helper
from backend.alerts_service import start_alerts_scheduler

# Initialize FastAPI app
app = FastAPI(
    title="PAC3220 Energy Monitoring API",
    description="Prepaid energy monitoring system for Siemens PAC3220 analyzers",
    version="1.0.0",
    docs_url="/docs",
    redoc_url="/redoc"
)

env = os.getenv("APP_ENV", "development")
allow_origins_env = os.getenv("ALLOW_ORIGINS", "http://localhost:3000,http://localhost:5173")
allow_origins = [o.strip() for o in allow_origins_env.split(",")] if allow_origins_env else []
allow_credentials = env != "production"
app.add_middleware(
    CORSMiddleware,
    allow_origins=allow_origins,
    allow_credentials=allow_credentials,
    allow_methods=["*"],
    allow_headers=["*"],
)

class RateLimitMiddleware(BaseHTTPMiddleware):
    def __init__(self, app):
        super().__init__(app)
        self.limits = {}
        self.window = int(os.getenv("GLOBAL_RATE_LIMIT_WINDOW_SECONDS", "60"))
        self.max_per_key = int(os.getenv("GLOBAL_RATE_LIMIT_PER_WINDOW", "300"))

    async def dispatch(self, request, call_next):
        try:
            ip = request.client.host if request.client else "unknown"
            key = f"{ip}:{request.url.path}"
            now = time()
            buf = self.limits.get(key, [])
            buf = [t for t in buf if now - t <= self.window]
            buf.append(now)
            self.limits[key] = buf
            if len(buf) > self.max_per_key:
                return JSONResponse(status_code=429, content={"detail": "Too many requests"})
        except Exception:
            pass
        return await call_next(request)

app.add_middleware(RateLimitMiddleware)

# Global error handlers with unified JSON and DB logging
@app.exception_handler(HTTPException)
async def http_exception_handler(request: Request, exc: HTTPException):
    try:
        db_helper.execute_query(
            """
            INSERT INTO ops.Events (Level, EventType, Message, Source, MetaData)
            VALUES ('WARN', 'http_exception', ?, 'API', ?)
            """,
            (str(exc.detail), f"{'{'}\"path\": \"{str(request.url.path)}\", \"status\": {exc.status_code}{'}'}"),
        )
    except Exception:
        pass
    return JSONResponse(status_code=exc.status_code, content={
        "success": False,
        "error": {
            "code": exc.status_code,
            "message": exc.detail,
            "path": str(request.url.path),
        }
    })

@app.exception_handler(Exception)
async def unhandled_exception_handler(request: Request, exc: Exception):
    try:
        db_helper.execute_query(
            """
            INSERT INTO ops.Events (Level, EventType, Message, Source, MetaData)
            VALUES ('ERROR', 'unhandled_exception', ?, 'API', ?)
            """,
            (str(exc), f"{'{'}\"path\": \"{str(request.url.path)}\"{'}'}"),
        )
    except Exception:
        pass
    return JSONResponse(status_code=500, content={
        "success": False,
        "error": {
            "code": 500,
            "message": "Internal server error",
            "path": str(request.url.path),
        }
    })

# Include routers
app.include_router(auth_router, prefix="/api", tags=["Authentication"])
app.include_router(admin_router, prefix="/api/admin", tags=["Admin"])
app.include_router(dashboard_router, prefix="/api/dashboard", tags=["Dashboard"])
app.include_router(devices_router, prefix="/api/devices", tags=["Devices"])
app.include_router(readings_router, prefix="/api/readings", tags=["Readings"])
app.include_router(tariffs_router, prefix="/api/tariffs", tags=["Tariffs"])
app.include_router(do_router, prefix="/api/admin/do", tags=["DigitalOutput"])

# WebSocket endpoints
@app.websocket("/ws/dashboard")
async def dashboard_websocket(websocket: WebSocket, user_id: int = None):
    """WebSocket endpoint for real-time dashboard updates"""
    await ws_manager.connect(websocket, "dashboard", user_id)
    try:
        while True:
            # Keep connection alive, data is pushed from server
            data = await websocket.receive_text()
            # Handle any client messages if needed
    except WebSocketDisconnect:
        ws_manager.disconnect(websocket, "dashboard", user_id)

@app.websocket("/ws/admin")
async def admin_websocket(websocket: WebSocket, user_id: int = None):
    """WebSocket endpoint for admin real-time updates"""
    await ws_manager.connect(websocket, "admin", user_id)
    try:
        while True:
            data = await websocket.receive_text()
    except WebSocketDisconnect:
        ws_manager.disconnect(websocket, "admin", user_id)

# Health check endpoint
@app.get("/health")
async def health_check():
    """Health check endpoint"""
    return {"status": "ok"}

# Root endpoint
@app.get("/")
async def root():
    """Root endpoint with system information"""
    return {
        "message": "PAC3220 Prepaid Energy Monitoring System API",
        "version": "1.0.0",
        "docs": "/docs",
        "health": "/health"
    }

# Startup: begin alerts scheduler if enabled
@app.on_event("startup")
async def startup_alerts():
    if os.getenv("ALERTS_ENABLED", "false").lower() == "true":
        try:
            await start_alerts_scheduler()
        except Exception:
            pass

# Startup: begin background tasks
@app.on_event("startup")
async def startup_event():
    """Initialize background tasks on startup"""
    import asyncio
    from backend.websocket_manager import periodic_status_updates

    # Optional disable via environment to avoid errors in dev without DB
    if os.getenv("DISABLE_BACKGROUND_TASKS", "false").lower() == "true":
        return
    # Start WebSocket status updates only if DB is reachable
    try:
        if db_helper.test_connection():
            asyncio.create_task(periodic_status_updates())
    except Exception:
        pass

    # Unified poller will run as an independent service/process.
    # API no longer launches legacy poller threads to avoid coupling and blocking.


if __name__ == "__main__":
    # Get configuration from environment
    host = os.getenv("API_HOST", "0.0.0.0")
    port = int(os.getenv("API_PORT", "8000"))
    reload = os.getenv("APP_ENV", "development") == "development"

    print("=" * 60)
    print("PAC3220 Prepaid Energy Monitoring System")
    print("=" * 60)
    print(f"[STARTING] API server on {host}:{port}")
    print(f"[DOCS] API Documentation: http://{host}:{port}/docs")
    print(f"[RELOAD] Reload enabled: {reload}")
    print("=" * 60)

    # Startup handlers are defined above

    uvicorn.run(
        "main:app",
        host=host,
        port=port,
        reload=reload,
        log_level="info"
    )
@app.post("/api/auth/login", response_model=TokenResponse)
async def proxy_login_auth(req: LoginRequest, http_req: Request):
    # Early dev fallback: allow admin login without DB before calling router
    try:
        if os.getenv("APP_ENV", "development").lower() == "development" and (req.username or "").lower() == "admin":
            token = create_jwt_token(user_id=1, username="admin", role="Admin")
            return TokenResponse(
                success=True,
                token=token,
                refresh_token=token,
                user={
                    "id": 1,
                    "username": "admin",
                    "fullname": "Administrator",
                    "email": "admin@example.com",
                    "role": "Admin",
                },
            )
    except Exception:
        pass
    try:
        return await auth_login(req, http_req)
    except Exception as e:
        try:
            # Development fallback to unblock login when DB is misconfigured
            if os.getenv("APP_ENV", "development").lower() == "development" and (req.username or "").lower() == "admin":
                token = create_jwt_token(user_id=1, username="admin", role="Admin")
                return TokenResponse(
                    success=True,
                    token=token,
                    refresh_token=token,
                    user={
                        "id": 1,
                        "username": "admin",
                        "fullname": "Administrator",
                        "email": "admin@example.com",
                        "role": "Admin",
                    },
                )
        except Exception:
            pass
        raise e
@app.post("/api/login", response_model=TokenResponse)
async def proxy_login(req: LoginRequest, http_req: Request):
    return await auth_login(req, http_req)

@app.get("/api/user/dashboard")
async def alias_user_dashboard(current_user: dict = Depends(get_current_user)):
    try:
        user_id = current_user.get("sub")
        # Try stored procedure first
        try:
            result = db_helper.execute_stored_procedure("app.sp_GetUserDashboard", {"@UserID": user_id})
        except Exception:
            try:
                result = db_helper.execute_query("SELECT * FROM app.vw_UserDashboard WHERE UserID = ?", (user_id,))
            except Exception:
                result = []
        return {"success": True, "data": result[0] if result else {}, "timestamp": datetime.utcnow()}
    except HTTPException:
        raise
    except Exception as e:
        print(f"Alias user dashboard error: {e}")
        return {"success": True, "data": {}, "timestamp": datetime.utcnow()}
