EnergyMonitoringSystem/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pyodbc==5.1.0
pymodbus==3.6.9
tenacity==9.0.0
python-jose==3.3.0
bcrypt==4.2.0
pydantic==2.9.2
email-validator==2.2.0
websockets==12.0
orjson==3.10.7
httpx==0.27.2
pytest==8.3.3
